# docs/openapi-generated.yaml
openapi: 3.1.0
info:
  title: 'Bar Assistant API'
...
  contact:
    name: 'Bar Assistant'
    url: 'https://barassistant.app'
    email: info@barassistant.app
  license:
    name: MIT
    url: 'https://github.com/karlomikus/bar-assistant/blob/master/LICENSE'
  version: '{{VERSION}}'
servers:
  -
    url: 'http://localhost:8000/api'
    description: 'Local docker development'
  -
    url: 'https://api.barassistant.app/api'
    description: Production
paths:
...
  /cocktails:
    get:
      tags:
        - Cocktails
      summary: 'List cocktails'
      description: 'Show a list of all cocktails in a bar'
      operationId: listCocktails
      parameters:
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
        -
          name: page
          in: query
          description: 'Set current page number'
          schema:
            type: integer
        -
          name: per_page
          in: query
          description: 'Set number of results per page'
          schema:
            type: integer
        -
          name: filter
          in: query
          description: 'Filter by attributes'
          style: deepObject
          explode: true
          schema:
            properties:
              id:
                type: string
              name:
                type: string
              ingredient_name:
                type: string
              tag_id:
                type: string
              created_user_id:
                type: string
              glass_id:
                type: string
              cocktail_method_id:
                type: string
              collection_id:
                type: string
              favorites:
                type: boolean
              on_shelf:
                type: boolean
              bar_shelf:
                type: boolean
              user_shelves:
                type: string
              shelf_ingredients:
                type: string
              is_public:
                type: boolean
              user_rating_min:
                type: string
              user_rating_max:
                type: string
              average_rating_min:
                type: string
              average_rating_max:
                type: string
              abv_min:
                type: string
              abv_max:
                type: string
              main_ingredient_id:
                type: string
              total_ingredients:
                type: string
              missing_ingredients:
                type: string
              specific_ingredients:
                type: string
              ignore_ingredients:
                type: string
            type: object
        -
          name: sort
          in: query
          description: 'Sort by attributes. Available attributes: `name`, `created_at`, `average_rating`, `user_rating`, `abv`, `total_ingredients`, `missing_ingredients`, `missing_bar_ingredients`, `favorited_at`.'
          schema:
            type: string
        -
          name: include
          in: query
          description: 'Include additional relationships. Available relations: `glass`, `method`, `user`, `navigation`, `utensils`, `createdUser`, `updatedUser`, `images`, `tags`, `ingredients.ingredient`, `ratings`.'
          schema:
            type: string
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Cocktail' } }
                  links: { properties: { first: { type: [string, 'null'] }, last: { type: [string, 'null'] }, prev: { type: [string, 'null'] }, next: { type: [string, 'null'] } }, type: object }
                  meta: { properties: { current_page: { type: integer }, from: { type: integer }, last_page: { type: integer }, links: { type: array, items: { properties: { url: { type: string }, label: { type: string }, active: { type: boolean } }, type: object } }, path: { type: string }, per_page: { type: integer }, to: { type: integer }, total: { type: integer } }, type: object }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
    post:
      tags:
        - Cocktails
      summary: 'Create cocktail'
      description: 'Create a new cocktail'
      operationId: saveCocktail
      parameters:
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CocktailRequest'
      responses:
        '201':
          description: 'Successful response'
          headers:
            Location:
              description: 'URL of the new resource'
              schema:
                type: string
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/Cocktail' }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '422':
          description: 'Request validation failed.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  /collections:
    get:
      tags:
        - Collections
      summary: 'List collections'
      description: 'Show a list of all user collections in a specific bar'
      operationId: listCollections
      parameters:
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
        -
          name: filter
          in: query
          description: 'Filter by attributes'
          style: deepObject
          explode: true
          schema:
            properties:
              id:
                type: integer
              name:
                type: string
              cocktail_id:
                type: string
            type: object
        -
          name: include
          in: query
          description: 'Include additional relationships. Available relations: `cocktails`.'
          schema:
            type: string
        -
          name: sort
          in: query
          description: 'Sort by attributes. Available attributes: `name`, `created_at`.'
          schema:
            type: string
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Collection' } }
                type: object
    post:
      tags:
        - Collections
      summary: 'Create collection'
      description: 'Create a new collection'
      operationId: saveCollection
      parameters:
        -
          name: Bar-Assistant-Bar-Id
          in: header
          description: 'Database id of a bar. Required if you are not using `bar_id` query string.'
          required: false
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionRequest'
      responses:
        '201':
          description: 'Successful response'
          headers:
            Location:
              description: 'URL of the new resource'
              schema:
                type: string
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/Collection' }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
  '/bars/{id}/collections':
    get:
      tags:
        - Collections
      summary: 'List shared collections'
      description: 'Show a list of all collections that users shared with the bar'
      operationId: listSharedCollections
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Collection' } }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
  '/collections/{id}':
    get:
      tags:
        - Collections
      summary: 'Show collection'
      description: 'Show a specific collection'
      operationId: showCollection
      parameters:
        -
          name: id
          in: path
          description: 'Database id or slug of a resource'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/Collection' }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
    put:
      tags:
        - Collections
      summary: 'Update collection'
      description: 'Update a specific collection'
      operationId: updateCollection
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionRequest'
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/Collection' }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
    delete:
      tags:
        - Collections
      summary: 'Delete collection'
      description: 'Delete a specific collection'
      operationId: deleteCollection
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Successful response'
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
        '404':
          description: 'Resource record not found.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
  '/collections/{id}/cocktails':
    put:
      tags:
        - Collections
      summary: 'Sync cocktails in a collection'
      description: 'Used to updated/add/delete cocktails in a collection. To delete all cocktails pass an empty array.'
      operationId: syncCocktailsInCollection
      parameters:
        -
          name: id
          in: path
          description: 'Database id of a resource'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                cocktails:
                  type: array
                  items: { type: integer }
              type: object
      responses:
        '200':
          description: 'Successful response'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                required:
                  - data
                properties:
                  data: { $ref: '#/components/schemas/Collection' }
                type: object
        '403':
          description: 'You are not authorized for this action.'
          headers:
            x-ratelimit-limit:
              description: 'Max number of attempts.'
              schema:
                type: integer
            x-ratelimit-remaining:
              description: 'Remaining number of attempts.'
              schema:
                type: integer
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/APIError' }
                type: object
...
components:
  schemas:
...
    APIError:
      required:
        - type
        - message
      properties:
        type:
          type: string
          example: api_error
        message:
          type: string
          example: 'Resource record not found.'
      type: object
...
    Cocktail:
      required:
        - id
        - name
        - slug
        - garnish
        - description
        - instructions
        - source
        - public_id
        - public_at
        - created_at
        - updated_at
        - abv
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Cocktail name'
        slug:
          type: string
          example: cocktail-name-1
        instructions:
          type: string
          example: 'Step by step instructions'
        garnish:
          type:
            - string
            - 'null'
          example: Garnish
        description:
          type:
            - string
            - 'null'
          example: 'Cocktail description'
        source:
          type:
            - string
            - 'null'
          example: 'Source of the recipe'
        public_id:
          type:
            - string
            - 'null'
          example: public-id-1
        public_at:
          type:
            - string
            - 'null'
          format: date-time
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        tags:
          type: array
          items:
            properties:
              id:
                type: integer
                example: 1
              name:
                type: string
                example: 'Tag name'
            type: object
        rating:
          required:
            - user
            - average
            - total_votes
          properties:
            user:
              description: "Current user's rating"
              type:
                - integer
                - 'null'
              example: 1
            average:
              description: 'Average rating'
              type: integer
              example: 4
            total_votes:
              type: integer
              example: 12
          type: object
        glass:
          oneOf:
            -
              $ref: '#/components/schemas/Glass'
            -
              type: 'null'
        utensils:
          type: array
          items:
            $ref: '#/components/schemas/Utensil'
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/CocktailIngredient'
        created_at:
          type: string
          format: date-time
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
        method:
          oneOf:
            -
              $ref: '#/components/schemas/CocktailMethod'
            -
              type: 'null'
        abv:
          type:
            - number
            - 'null'
          format: float
          example: 40
        volume_ml:
          type:
            - number
            - 'null'
          format: float
          example: 67.5
        alcohol_units:
          type:
            - number
            - 'null'
          format: float
          example: 25.5
        calories:
          type:
            - integer
            - 'null'
          example: 350
        created_user:
          $ref: '#/components/schemas/UserBasic'
        updated_user:
          oneOf:
            -
              $ref: '#/components/schemas/UserBasic'
            -
              type: 'null'
        in_shelf:
          type: boolean
        in_bar_shelf:
          type: boolean
        access:
          required:
            - can_edit
            - can_delete
            - can_rate
            - can_add_note
          properties:
            can_edit:
              type: boolean
              example: true
            can_delete:
              type: boolean
              example: true
            can_rate:
              type: boolean
              example: true
            can_add_note:
              type: boolean
              example: true
          type: object
        navigation:
          required:
            - prev
            - next
          properties:
            prev:
              type:
                - string
                - 'null'
              example: old-fashioned-1
            next:
              type:
                - string
                - 'null'
              example: tom-collins-1
          type: object
      type: object
...
    Utensil:
      required:
        - id
        - name
        - description
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Shaker
        description:
          type:
            - string
            - 'null'
          example: 'Used to shake ingredients'
      type: object
...
    ValidationError:
      required:
        - errors
        - message
      properties:
        message:
          type: string
          example: 'The cocktail name must be a string. (and 2 more errors)'
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      type: object
  securitySchemes:
    user_token:
      type: http
      scheme: bearer
security:
  -
    user_token: []
...